requirements.txt
# Core dependencies
streamlit>=1.28.0
psycopg2-binary>=2.9.0
pgvector>=0.2.0
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0

# NLP and ML
sentence-transformers>=2.2.0
transformers>=4.30.0
torch>=2.0.0
scikit-learn>=1.3.0
nltk>=3.8.0

# Document processing
PyPDF2>=3.0.0
python-docx>=1.0.0
beautifulsoup4>=4.12.0
requests>=2.31.0

# Evaluation and visualization
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.15.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.65.0
pydantic>=2.0.0
typer>=0.9.0

# Database
sqlalchemy>=2.0.0

# Optional: For better sparse vectors
rank-bm25>=0.2.0

# New dependencies for advanced features
textstat>=0.7.0  # For readability metrics
wordcloud>=1.9.0  # For topic visualization
matplotlib-venn>=0.11.0  # For set visualizations
